requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
aiosqlite==0.17.0
astunparse==1.6.3
async-timeout==4.0.2
asyncio==3.4.3
asyncio-dgram==2.1.2
asyncpg==0.26.0
attrs==21.4.0
black==22.6.0
braceexpand==0.1.7
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
discord.py @ git+https://github.com/Rapptz/discord.py@1fa7d7e40201b8a0eaee10a6b51786c60bc9d63f
dnspython==2.2.1
dnspython3==1.15.0
frozenlist==1.3.0
fuzzywuzzy==0.18.0
hiredis==2.1.0
idna==3.3
import-expression==1.1.4
importlib-metadata==4.12.0
jishaku==2.5.1
line-profiler==3.5.1
motor==3.0.0
multidict==6.0.2
mypy-extensions==0.4.3
networkx==2.8.8
parsedatetime==2.6
pathspec==0.9.0
platformdirs==2.5.2
pycryptodome==3.15.0
pymongo==4.2.0
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
yarl==1.7.2
zipp==3.8.1
